+++ eland/ml/ml_model.py
@@ -20,7 +20,7 @@ from typing import TYPE_CHECKING, Any, Dict, List, Mapping, Optional, Tuple, Uni
 import elasticsearch
 import numpy as np
 
-from eland.common import ensure_es_client, es_version
+from eland.common import ensure_es_client, es_version, is_serverless_es
 from eland.utils import deprecated_api
 
 from .common import TYPE_CLASSIFICATION, TYPE_LEARNING_TO_RANK, TYPE_REGRESSION
@@ -504,7 +504,9 @@ class MLModel:
         )
         serializer = transformer.transform()
         model_type = transformer.model_type
-        default_inference_config: Mapping[str, Mapping[str, Any]] = {model_type: {}}
+
+        if inference_config is None:
+            inference_config = {model_type: {}}
 
         if es_if_exists is None:
             es_if_exists = "fail"
@@ -523,18 +525,25 @@ class MLModel:
         elif es_if_exists == "replace":
             ml_model.delete_model()
 
+        trained_model_input = None
+        is_ltr = next(iter(inference_config)) is TYPE_LEARNING_TO_RANK
+        if not is_ltr or (
+            es_version(es_client) < (8, 15) and not is_serverless_es(es_client)
+        ):
+            trained_model_input = {"field_names": feature_names}
+
         if es_compress_model_definition:
             ml_model._client.ml.put_trained_model(
                 model_id=model_id,
-                input={"field_names": feature_names},
-                inference_config=inference_config or default_inference_config,
+                inference_config=inference_config,
+                input=trained_model_input,
                 compressed_definition=serializer.serialize_and_compress_model(),
             )
         else:
             ml_model._client.ml.put_trained_model(
                 model_id=model_id,
-                input={"field_names": feature_names},
-                inference_config=inference_config or default_inference_config,
+                inference_config=inference_config,
+                input=trained_model_input,
                 definition=serializer.serialize_model(),
             )