+++ tests/__init__.py
@@ -20,7 +20,7 @@ import os
 import pandas as pd
 from elasticsearch import Elasticsearch
 
-from eland.common import es_version
+from eland.common import es_version, is_serverless_es
 
 ROOT_DIR = os.path.dirname(os.path.abspath(__file__))
 
@@ -33,6 +33,7 @@ ELASTICSEARCH_HOST = os.environ.get(
 ES_TEST_CLIENT = Elasticsearch(ELASTICSEARCH_HOST)
 
 ES_VERSION = es_version(ES_TEST_CLIENT)
+ES_IS_SERVERLESS = is_serverless_es(ES_TEST_CLIENT)
 
 FLIGHTS_INDEX_NAME = "flights"
 FLIGHTS_MAPPING = {
@@ -43,7 +44,7 @@ FLIGHTS_MAPPING = {
             "Carrier": {"type": "keyword"},
             "Dest": {"type": "keyword"},
             "DestAirportID": {"type": "keyword"},
-            "DestCityName": {"type": "keyword"},
+            "DestCityName": {"type": "keyword", "copy_to": "Cities"},
             "DestCountry": {"type": "keyword"},
             "DestLocation": {"type": "geo_point"},
             "DestRegion": {"type": "keyword"},
@@ -58,11 +59,12 @@ FLIGHTS_MAPPING = {
             "FlightTimeMin": {"type": "float"},
             "Origin": {"type": "keyword"},
             "OriginAirportID": {"type": "keyword"},
-            "OriginCityName": {"type": "keyword"},
+            "OriginCityName": {"type": "keyword", "copy_to": "Cities"},
             "OriginCountry": {"type": "keyword"},
             "OriginLocation": {"type": "geo_point"},
             "OriginRegion": {"type": "keyword"},
             "OriginWeather": {"type": "keyword"},
+            "Cities": {"type": "text"},
             "dayOfWeek": {"type": "byte"},
             "timestamp": {"type": "date", "format": "strict_date_hour_minute_second"},
         }